server/requirements.txt
Flask==2.3.3
websockets==11.0.3
Pillow==10.0.1
google-generativeai==0.3.2
asyncio
keyboard